requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn

tensorflow
keras

streamlit
Flask==2.1.1
Pillow==8.4.0

pickle-mixin==1.0.2
Werkzeug==2.0.1
gunicorn==20.1.0


mlflow

fastapi
uvicorn
python-multipart

mlflow==1.20.2


